Move JakartaEE repeat action to FAT infra
