+++ dev/build.example_fat/build.gradle
@@ -11,10 +11,7 @@
 
 // Define G:A:V coordinates of each dependency
 dependencies {
-  requiredLibs 'commons-logging:commons-logging:1.1.3',
-        'biz.aQute.bnd:biz.aQute.bndlib:4.3.1',
-        'commons-cli:commons-cli:1.4',
-        'org.slf4j:slf4j-simple:1.7.29',
-        'org.slf4j:slf4j-api:1.7.26',
-        'com.ibm.ws:jakartaee-transformer:0.0.5'
+  requiredLibs 'commons-logging:commons-logging:1.1.3'
 }
+
+addRequiredLibraries.dependsOn addJakartaTransformer
