+++ dev/build.example_fat/fat/src/com/ibm/ws/example/FATSuite.java
@@ -16,6 +16,7 @@ import org.junit.runners.Suite;
 import org.junit.runners.Suite.SuiteClasses;
 
 import componenttest.rules.repeater.FeatureReplacementAction;
+import componenttest.rules.repeater.JakartaEE9Action;
 import componenttest.rules.repeater.RepeatTests;
 
 @RunWith(Suite.class)
@@ -28,6 +29,6 @@ public class FATSuite {
     // First without any modifications, then again with all features upgraded to their EE8 equivalents.
     @ClassRule
     public static RepeatTests r = RepeatTests.with(FeatureReplacementAction.EE8_FEATURES())
-                    .andWith(new JakartaEEAction());
+                    .andWith(new JakartaEE9Action());
 
 }
