+++ dev/fattest.simplicity/bnd.bnd
@@ -59,6 +59,7 @@ test.project: true
 	commons-httpclient:commons-httpclient;version=3.1,\
 	com.ibm.websphere.javaee.jsonp.1.0;version=latest,\
 	com.ibm.websphere.javaee.servlet.3.0;version=latest,\
+    com.ibm.ws:jakartaee-transformer;version=0.0.5,\
 	com.ibm.ws.common.encoder;version=latest,\
 	com.ibm.ws.componenttest;version=latest,\
 	com.ibm.ws.componenttest:com.ibm.componenttest.common;version=1.0.0,\
